+++ Solver/src/VCellFront.cpp
@@ -313,6 +313,6 @@ std::vector<std::vector<spatial::TPoint<FCT,2> > > VCellFront<FCT>::retrieveCurv
 }
 
 template class VCellFront<double>;
-template class VCellFront<int>;
-template class VCellFront<long>;
+template class VCellFront<int16_t>;
+template class VCellFront<int32_t>;
 template class VCellFront<int64_t>;
